+++ src/backend/joanie/tests/core/test_api_admin_order_group.py
@@ -2,10 +2,13 @@
 Test suite for OrderGroup Admin API.
 """
 
+from datetime import timedelta
 from http import HTTPStatus
 from operator import itemgetter
 
+from django.db import IntegrityError
 from django.test import TestCase
+from django.utils import timezone as django_timezone
 
 from joanie.core import factories, models
 
@@ -53,10 +56,13 @@ class OrderGroupAdminApiTest(TestCase):
                 "id": str(order_group.id),
                 "nb_seats": order_group.nb_seats,
                 "is_active": order_group.is_active,
+                "is_enabled": order_group.is_enabled,
                 "nb_available_seats": order_group.nb_seats
                 - order_group.get_nb_binding_orders(),
                 "created_on": order_group.created_on.isoformat().replace("+00:00", "Z"),
                 "can_edit": True,
+                "start": None,
+                "end": None,
             }
             for order_group in order_groups
         ]
@@ -121,17 +127,20 @@ class OrderGroupAdminApiTest(TestCase):
             "id": str(order_group.id),
             "nb_seats": order_group.nb_seats,
             "is_active": order_group.is_active,
+            "is_enabled": order_group.is_enabled,
             "nb_available_seats": order_group.nb_seats
             - order_group.get_nb_binding_orders(),
             "created_on": order_group.created_on.isoformat().replace("+00:00", "Z"),
             "can_edit": True,
+            "start": None,
+            "end": None,
         }
         self.assertEqual(content, expected_return)
 
     # create
     def test_admin_api_order_group_create_anonymous(self):
         """
-        Anonymous users should not be able to create an order groups.
+        Anonymous users should not be able to create an order group.
         """
         relation = factories.CourseProductRelationFactory()
 
@@ -146,6 +155,33 @@ class OrderGroupAdminApiTest(TestCase):
         self.assertEqual(
             content["detail"], "Authentication credentials were not provided."
         )
+
+    def test_admin_api_order_group_create_authenticated_with_nb_seats_is_none(self):
+        """
+        Authenticated users should be able to create an order group and set None for
+        `nb_seats`.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+        relation = factories.CourseProductRelationFactory()
+        data = {
+            "nb_seats": None,
+            "is_active": True,
+        }
+
+        response = self.client.post(
+            f"{self.base_url}/{relation.id}/order-groups/",
+            content_type="application/json",
+            data=data,
+        )
+
+        content = response.json()
+
+        self.assertEqual(response.status_code, HTTPStatus.CREATED)
+        self.assertIsNone(content["nb_seats"])
+        self.assertEqual(content["is_active"], data["is_active"])
+        self.assertTrue(content["is_enabled"])
+        self.assertEqual(models.OrderGroup.objects.filter(**data).count(), 1)
 
     def test_admin_api_order_group_create_authenticated(self):
         """
@@ -312,5 +348,231 @@ class OrderGroupAdminApiTest(TestCase):
             response = self.client.delete(
                 f"{self.base_url}/{relation.id}/order-groups/{order_group.id}/",
             )
+
         self.assertEqual(response.status_code, HTTPStatus.NO_CONTENT)
         self.assertFalse(models.OrderGroup.objects.filter(id=order_group.id).exists())
+
+    def test_admin_api_order_group_create_start_and_end_date(self):
+        """
+        Authenticated admin user should be able to create an order group and set
+        a start and end date.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+
+        relation = factories.CourseProductRelationFactory()
+        data = {
+            "start": "2025-06-01T00:00:00Z",
+            "end": "2025-06-20T00:00:00Z",
+            "nb_seats": 10,
+            "is_active": False,
+        }
+
+        response = self.client.post(
+            f"{self.base_url}/{relation.id}/order-groups/",
+            content_type="application/json",
+            data=data,
+        )
+
+        self.assertEqual(response.status_code, HTTPStatus.CREATED)
+
+        content = response.json()
+
+        self.assertEqual(content["start"], data["start"])
+        self.assertEqual(content["end"], data["end"])
+        self.assertFalse(content["is_active"])
+
+    def test_admin_api_order_group_create_start_date_greater_than_end_date(self):
+        """
+        Authenticated admin user should not be able to create an order group when the
+        start date is greater than the end date.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+
+        relation = factories.CourseProductRelationFactory()
+        data = {
+            "start": "2025-01-20T00:00:00Z",
+            "end": "2025-01-01T00:00:00Z",
+        }
+
+        with self.assertRaises(IntegrityError):
+            self.client.post(
+                f"{self.base_url}/{relation.id}/order-groups/",
+                content_type="application/json",
+                data=data,
+            )
+
+    def test_admin_api_order_group_update_start_and_end_date(self):
+        """
+        Authenticated admin user can update the start and end date of an existing
+        order group.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+
+        relation = factories.CourseProductRelationFactory()
+        order_group = factories.OrderGroupFactory(
+            course_product_relation=relation,
+            start="2025-01-11T00:00:00Z",
+            end="2025-01-20T00:00:00Z",
+        )
+
+        data = {
+            "start": "2025-02-13T00:00:00Z",
+            "end": "2025-02-19T00:00:00Z",
+        }
+
+        response = self.client.put(
+            f"{self.base_url}/{relation.id}/order-groups/{order_group.id}/",
+            content_type="application/json",
+            data=data,
+        )
+
+        content = response.json()
+
+        self.assertEqual(response.status_code, HTTPStatus.OK)
+        self.assertEqual(content["start"], data["start"])
+        self.assertEqual(content["end"], data["end"])
+
+    def test_admin_api_order_group_is_enabled_and_is_active(
+        self,
+    ):
+        """
+        When the order group is not yet active, even if the dates qualifies for early-birds
+        or last minutes sales, the property `is_enabled` should return False. Otherwise,
+        when the order group is active, the computed value of the property `is_enabled` will
+        return True if the datetimes meet the conditions.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+
+        relation = factories.CourseProductRelationFactory()
+        test_cases = [
+            {
+                "start": None,
+                "end": django_timezone.now() + timedelta(days=1),
+            },
+            {"start": django_timezone.now(), "end": None},
+            {
+                "start": django_timezone.now() - timedelta(days=1),
+                "end": django_timezone.now() + timedelta(days=1),
+            },
+        ]
+
+        for case in test_cases:
+            with self.subTest(start=case["start"], end=case["end"]):
+                order_group = factories.OrderGroupFactory(
+                    course_product_relation=relation,
+                    is_active=False,
+                    start=case["start"],
+                    end=case["end"],
+                )
+
+                response = self.client.get(
+                    f"{self.base_url}/{relation.id}/order-groups/{order_group.id}/"
+                )
+
+                content = response.json()
+
+                self.assertEqual(response.status_code, HTTPStatus.OK)
+                self.assertFalse(content["is_active"])
+                self.assertFalse(content["is_enabled"])
+
+                order_group.is_active = True
+                order_group.save()
+
+                response = self.client.get(
+                    f"{self.base_url}/{relation.id}/order-groups/{order_group.id}/"
+                )
+
+                content = response.json()
+
+                self.assertEqual(response.status_code, HTTPStatus.OK)
+                self.assertTrue(content["is_active"])
+                self.assertTrue(content["is_enabled"])
+
+    def test_admin_api_order_group_is_not_enabled_start_end_outside_datetime_range(
+        self,
+    ):
+        """
+        When the order group is active but is not yet within the datetime ranges to be enabled,
+        it should return False.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+
+        relation = factories.CourseProductRelationFactory()
+        test_cases = [
+            {
+                "start": None,
+                "end": django_timezone.now() - timedelta(days=1),
+            },
+            {"start": django_timezone.now() + timedelta(days=1), "end": None},
+            {
+                "start": django_timezone.now() + timedelta(days=1),
+                "end": django_timezone.now() + timedelta(days=2),
+            },
+        ]
+
+        for case in test_cases:
+            with self.subTest(start=case["start"], end=case["end"]):
+                order_group = factories.OrderGroupFactory(
+                    course_product_relation=relation,
+                    is_active=True,
+                    start=case["start"],
+                    end=case["end"],
+                )
+
+                response = self.client.get(
+                    f"{self.base_url}/{relation.id}/order-groups/{order_group.id}/"
+                )
+
+                content = response.json()
+
+                self.assertEqual(response.status_code, HTTPStatus.OK)
+                self.assertTrue(content["is_active"])
+                self.assertFalse(content["is_enabled"])
+
+    def test_admin_api_order_group_is_active_and_nb_seats_is_enabled(self):
+        """
+        When the order group is active and the number of seats is None,
+        `is_enabled` should return the value True. Otherwise, it should
+        return False.
+        """
+        admin = factories.UserFactory(is_staff=True, is_superuser=True)
+        self.client.login(username=admin.username, password="password")
+
+        relation = factories.CourseProductRelationFactory()
+        order_group = factories.OrderGroupFactory(
+            course_product_relation=relation,
+            is_active=True,
+            nb_seats=None,
+            start=None,
+            end=None,
+        )
+
+        response = self.client.get(
+            f"{self.base_url}/{relation.id}/order-groups/{order_group.id}/"
+        )
+
+        content = response.json()
+
+        self.assertTrue(content["is_enabled"])
+        self.assertTrue(content["is_active"])
+        self.assertIsNone(content["nb_available_seats"])
+        self.assertIsNone(content["nb_seats"])
+
+        order_group.is_active = False
+        order_group.save()
+
+        response = self.client.get(
+            f"{self.base_url}/{relation.id}/order-groups/{order_group.id}/"
+        )
+
+        content = response.json()
+
+        self.assertFalse(content["is_enabled"])
+        self.assertFalse(content["is_active"])
+        self.assertIsNone(content["nb_available_seats"])
+        self.assertIsNone(content["nb_seats"])